Refactor i18n structure: move to features-based architecture
- Move i18n files from src/i18n/ to src/features/i18n/
- Update next.config.ts to point to new request.ts location
- Update language-switcher imports to use new paths
- Maintain all existing functionality and tests
